requirements.txt
# Core Data Science
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0

# ML Models
lightgbm>=4.0.0
xgboost>=1.7.0

# NLP & Transformers
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
datasets==3.6.0

# Financial ML
yfinance>=0.2.0
pandas-datareader>=0.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Model Explainability
# shap>=0.41.0  # Temporarily disabled for Python 3.11+ compatibility

# API (Optional)
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Development & Testing
pytest>=7.4.0
jupyter>=1.0.0
jupyterlab>=4.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Environment & Config
python-dotenv>=1.0.0
pyyaml>=6.0

# RAG and Vector Store
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
openai>=1.0.0

# Model Explainability  
shap>=0.42.0

# Financial NLP and Feature Engineering
vaderSentiment>=3.3.0
textstat>=0.7.0
nltk>=3.8.0
scikit-learn>=1.3.0
